requirements.txt
openai
python-dotenv
langchain
langchain[openai]
neo4j 
pandas
deepeval
sentence-transformers
pip install faiss-cpu
langchain-community